requirements.txt
django==4.2.6
python-decouple==3.5
django-ninja==0.22.1
django-ninja-extra==0.18.9
django-tables2==2.6.0
requests==2.31.0
django-crispy-forms==2.0
crispy-bootstrap4
django_select2
qrcode[pil]